[FIX] local path 경로 찾을 수 없음 이슈 해결 - 비디오의 경우 s3 url 기반 다운로드로 수정 단, local path 있는지 확인 후 있으면 로컬 이용 - os.remove()의 경우 있을 때만 삭제되도록 수정정

+++ clip_video_info.py
@@ -4,12 +4,35 @@ from s3_client import s3_client
 from botocore.exceptions import NoCredentialsError
 from drama_crawling import search_drama, get_drama
 from fastapi import HTTPException
+import re
 
 # S3 설정
 AWS_ACCESS_KEY = os.getenv("AWS_ACCESS_KEY")
 AWS_SECRET_KEY = os.getenv("AWS_SECRET_KEY")
 S3_BUCKET_NAME = os.getenv("S3_BUCKET_NAME", "test-fastapi-bucket")
 S3_REGION_NAME = os.getenv("S3_REGION_NAME", "ap-northeast-2")
+
+def parse_s3_url(s3_url: str):
+    regex = r"https://([^/]+)\.s3\.[^/]+\.amazonaws\.com/(.+)"
+    match = re.match(regex, s3_url)
+    if not match:
+        raise ValueError(f"Invalid S3 URL: {s3_url}")
+    return match.group(1), match.group(2)
+
+def get_video_from_s3(s3_url):
+    TEMP_DIR = "tmp"
+    bucket_name, object_key = parse_s3_url(s3_url)
+    local_path = os.path.join(TEMP_DIR, object_key.split('/')[-1])  # 임시 파일 경로 설정
+    print(f"Downloading video from S3 to {local_path}")
+
+    try:
+        s3_client.download_file(bucket_name, object_key, local_path)
+        print(f"Downloaded {local_path}")
+    except NoCredentialsError:
+        raise Exception("AWS credentials not available.")
+    except Exception as e:
+        raise Exception(f"Error downloading from S3: {e}")
+    
 
 def search_drama_api(drama_title: str):
     result = search_drama(drama_title)
@@ -63,7 +86,7 @@ def crop_and_pad_to_1080x1920(clip):
     
     return clip
 
-def insert_title_into_video(local_path, task_id, title, idx):
+def insert_title_into_video(local_path, task_id, title, idx, s3_url):
     TEMP_DIR = 'tmp'
     if not os.path.exists(TEMP_DIR):
         os.makedirs(TEMP_DIR)
@@ -71,13 +94,20 @@ def insert_title_into_video(local_path, task_id, title, idx):
     font_path = "/System/Library/Fonts/AppleSDGothicNeo.ttc"
 
     try:
+        if os.path.exists(local_path):
+            print(f"Using local video file: {local_path}")
+        else:
+            print(f"Local file not found, downloading from S3: {s3_url}")
+            get_video_from_s3(s3_url)  # S3에서 다운로드
+
         video = VideoFileClip(local_path)
 
         text = title.encode('utf-8')
         txt_clip = TextClip(text, fontsize=60, color='white', font=font_path)
-        txt_clip.save_frame("img_clip_title.png", t=0)
+        txt_clip_path = os.path.join(TEMP_DIR, "img_clip_title.png")
+        txt_clip.save_frame(txt_clip_path, t=0)
 
-        text_img = ImageClip("img_clip_title.png").set_duration(video.duration).set_position(('center', 300))
+        text_img = ImageClip(txt_clip_path).set_duration(video.duration).set_position(('center', 300))
 
         filename = f"{str(task_id)}_highlight_with_ai_title_{str(idx)}.mp4"
         output_path = os.path.join(TEMP_DIR, filename)
@@ -87,7 +117,7 @@ def insert_title_into_video(local_path, task_id, title, idx):
         
         s3_url_ = upload_to_s3(output_path, filename)
 
-        os.remove("img_clip_title.png")
+        os.remove(txt_clip_path)
 
         return s3_url_
     
@@ -96,7 +126,7 @@ def insert_title_into_video(local_path, task_id, title, idx):
     except Exception as e:
         return f"🚨 예기치 않은 오류 발생: {e}"
 
-def clip_and_save_highlights(local_path, task_id, drama_title, adjusted_highlights):
+def clip_and_save_highlights(local_path, task_id, drama_title, adjusted_highlights, s3_url):
     TEMP_DIR = 'tmp'
     if not os.path.exists(TEMP_DIR):
         os.makedirs(TEMP_DIR)
@@ -104,6 +134,12 @@ def clip_and_save_highlights(local_path, task_id, drama_title, adjusted_highligh
     font_path = "/System/Library/Fonts/AppleSDGothicNeo.ttc"
 
     try:
+        if os.path.exists(local_path):
+            print(f"Using local video file: {local_path}")
+        else:
+            print(f"Local file not found, downloading from S3: {s3_url}")
+            get_video_from_s3(s3_url)  # S3에서 다운로드
+
         video = VideoFileClip(local_path)
         
         drama_info = get_drama_api(drama_title)
@@ -114,11 +150,13 @@ def clip_and_save_highlights(local_path, task_id, drama_title, adjusted_highligh
 
         text1 = drama_title.encode('utf-8')
         txt_clip1 = TextClip(text1, fontsize=55, color='white', font=font_path)
-        txt_clip1.save_frame("img_clip1.png", t=0)
+        txt_img1_path = os.path.join(TEMP_DIR, "img_clip1.png")
+        txt_clip1.save_frame(txt_img1_path, t=0)
 
         text2 = f"{drama_info['broadcaster']} - {drama_info['air_date']}"
         txt_clip2 = TextClip(text2, fontsize=30, color='white', font=font_path)
-        txt_clip2.save_frame("img_clip2.png", t=0)
+        txt_img2_path = os.path.join(TEMP_DIR, "img_clip2.png")
+        txt_clip2.save_frame(txt_img2_path, t=0)
 
         url_list, output_path_list = [], []
         for idx, (start, end) in enumerate(adjusted_highlights):
@@ -129,8 +167,8 @@ def clip_and_save_highlights(local_path, task_id, drama_title, adjusted_highligh
             filename = f"{task_id}_highlight_with_info_{idx+1}.mp4"
             output_path = os.path.join(TEMP_DIR, filename)  # 임시 파일 경로 설정
             
-            text_img1 = ImageClip("img_clip1.png").set_duration(highlight_clip.duration).set_position(('center', highlight_clip.h - 450))
-            text_img2 = ImageClip("img_clip2.png").set_duration(highlight_clip.duration).set_position(('center', highlight_clip.h - 350))
+            text_img1 = ImageClip(txt_img1_path).set_duration(highlight_clip.duration).set_position(('center', highlight_clip.h - 450))
+            text_img2 = ImageClip(txt_img2_path).set_duration(highlight_clip.duration).set_position(('center', highlight_clip.h - 350))
         
             result = CompositeVideoClip([highlight_clip, text_img1, text_img2])
             # result.write_videofile(output_path, codec="libx264", audio_codec="aac", preset="ultrafast")
@@ -142,11 +180,12 @@ def clip_and_save_highlights(local_path, task_id, drama_title, adjusted_highligh
             url_list.append(s3_url_)
 
             print(f"Successfully saved highlight {idx + 1} in S3 BUCKET!!")
+            
+            if os.path.exists(local_path):
+                os.remove(local_path)
 
-            os.remove(local_path)
-
-        os.remove("img_clip1.png")
-        os.remove("img_clip2.png")
+        os.remove(txt_img1_path)
+        os.remove(txt_img2_path)
 
         return url_list
